Add messages and room messages auth, needs some testing

+++ src/ban/ban.service.ts
@@ -1,5 +1,7 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { Room } from 'src/room/entities/room.entity';
+import { RoomService } from 'src/room/room.service';
 import { Repository } from 'typeorm';
 import { CreateBanDto } from './dto/create-ban.dto';
 import { UpdateBanDto } from './dto/update-ban.dto';
@@ -10,14 +12,22 @@ export class BanService {
 
 	constructor(
 		@InjectRepository(Ban)
-		private bansRepository: Repository<Ban>,
+		private bansRepository: Repository<Ban>
 	) {}
 
-	async create(createBanDto: CreateBanDto) {
+
+	async create(sessionId: number, roomData: Room, createBanDto: CreateBanDto) {
+
+		// Get room owner ==> user_id != owner_id (you can't ban room's owner)
+		if(roomData.admins.includes(sessionId))
+			throw new HttpException({ message: 'You\'re not an admin of this room!' }, HttpStatus.UNAUTHORIZED);
+
+		if(createBanDto.user_id == roomData.owner_id)
+			throw new HttpException({ message: 'You can\'t ban the room crater!' }, HttpStatus.UNAUTHORIZED);
+
 		const newBannedUser = this.bansRepository.create(createBanDto); // or create({ ....my data });
 		
 		const data = await this.bansRepository.save(newBannedUser); // insert Or update if it already exists
-		console.log(data);
 
 		return data;
 		
@@ -27,14 +37,25 @@ export class BanService {
 		return this.bansRepository.find();
 	}
 
-	findUserInRoom(roomId: number, userId: number) {
-		return this.bansRepository.findOne({
+	async findUserInRoom(roomId: number, userId: number) {
+		const data = await this.bansRepository.findOne({
 			room_id: roomId,
 			user_id: userId,
 		});
+		if (!data)
+			throw new HttpException({ error: 'User Not Found' }, HttpStatus.NOT_FOUND);
+
+		return data;
 	}
 
-	async update(updateBanDto: UpdateBanDto) {
+	async update(sessionId: number, roomData: Room, updateBanDto: UpdateBanDto) {
+
+		if(roomData.admins.includes(sessionId))
+			throw new HttpException({ message: 'You\'re not an admin of this room!' }, HttpStatus.UNAUTHORIZED);
+
+		if(updateBanDto.user_id == roomData.owner_id)
+			throw new HttpException({ message: 'You can\'t ban the room crater' }, HttpStatus.UNAUTHORIZED);
+
 		const bannedUser = await this.findUserInRoom(updateBanDto.room_id, updateBanDto.user_id);
 
 		if( bannedUser )
@@ -42,24 +63,26 @@ export class BanService {
 			bannedUser.banned = updateBanDto.banned;
 			bannedUser.duration = updateBanDto.duration;
 			
-			await this.bansRepository.save(bannedUser);
-			return true;
+			return this.bansRepository.save(bannedUser);
 		}
 		{
-			return false;
+			throw new HttpException({ error: "Couldn't update, please try again later!" }, HttpStatus.NOT_MODIFIED);
 		}
 	}
 
-	async unbanUserFromRoom(roomId: number, userId: number) {
+	async unbanUserFromRoom(sessionId: number, roomData: Room , roomId: number, userId: number) {
+
+		if(roomData.admins.includes(sessionId))
+			throw new HttpException({ message: 'You\'re not an admin of this room!' }, HttpStatus.UNAUTHORIZED);
+
 		const unBannedUser = await this.findUserInRoom(roomId, userId);
 		if(unBannedUser)
 		{
-			await this.bansRepository.remove(unBannedUser);
-			return true;
+			return this.bansRepository.remove(unBannedUser);
 		}
 		else
 		{
-			return false
+			throw new HttpException({ error: "Couldn't update, please try again later!" }, HttpStatus.NOT_MODIFIED);
 		}
 	}
 }
